Refactor, move functional extensions.
