+++ src/main/java/jp/aha/oretama/typoChecker/TypoCheckerService.java
@@ -5,7 +5,7 @@ import jp.aha.oretama.typoChecker.model.Event;
 import jp.aha.oretama.typoChecker.model.Suggestion;
 import jp.aha.oretama.typoChecker.model.Token;
 import jp.aha.oretama.typoChecker.repository.GitHubRepository;
-import jp.aha.oretama.typoChecker.utils.FuntionalExtension;
+import jp.aha.oretama.typoChecker.utils.extenstions.FuntionalExtension;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.IOUtils;
